Create Assessment 1 (Part 1)
The aim of this exercise is to interpret a sequence of instructions for a very simple robot, and return the predicted state of the robot after the execution of the sequence of instructions. In effect, you are writing an interpreter for a very simple language.

The language is designed to guide a simple robot. The robot moves on an  grid, where the grid coordinates are integer values. The robot always starts at grid coordinates . The robot also has a direction: 'N', 'S', 'E' or 'W'. When the robot is at  and is facing in the direction 'N', then if it moves it will arrive at the location . Similarly, if the robot is at  and is facing in the direction 'E', then if it moves it will arrive at the location . The robot can peform two possible actions: it can move forwards one step to the next grid coordinate in the direction it is currently facing, or it can turn 90 degrees to the right (i.e., clockwise) on the spot. It is a very simple robot, as it can only move one step at a time, and can only turn to the right by 90 degrees at a time.

Our language allows us to describe a sequence of actions that the robot should perform in terms of the operators 'move' and 'turn', which correspond to the robot actions move and turn. At any point in time, the robot's state can be specified by a tuple (x, y, direction), where x and y specify the integer coordinates of the robot, and direction is the string specifying the direction that the robot is facing. For example, the initial state of the robot is always (0, 0, 'N'). After executing the operator 'move' from the initial state, the new state of the robot would be (0, 1, 'N'). Similarly, after executing the operator 'turn' from the initial state, the new state of the robot would be (0, 0, 'E').

A "program" in our language is given as a list of these two operators. Our "interpreter" for the language should be able to determine the new state of the robot, having executed the given program from the initial state of the robot. Here are three examples:

the program ['move', 'turn', 'move'] in our language would leave the robot in the state (1, 1, 'E')
the program ['move', 'turn', 'turn', 'move'] in our language would leave the robot in the state (0, 0, 'S')
the program ['turn', 'turn', 'turn', 'move', 'move'] in our language would leave the robot in the state (-2, 0, 'W').
In this exercise, you need to write a function simple_interpreter(code) that takes a program code in our language in the form of a list of strings, where each string is an operator in our robot language, and returns a tuple corresponding to the new state of the robot after executing the program. Note that the initial state is always (0, 0, 'N').

For your function, you can assume that the input is well formed; that is, the input list will always be a list of strings, and the strings will only be the operators 'move' or 'turn'. You will not be given lists containing any other values.

>>> print(simple_interpreter(['move', 'turn', 'move']))
(1, 1, 'E')
>>> print(simple_interpreter(['move', 'turn', 'turn', 'move']))
(0, 0, 'S')
>>> print(simple_interpreter(['turn', 'turn', 'turn', 'move', 'move']))
(-2, 0, 'W')
>>> print(simple_interpreter(['move', 'turn', 'move', 'turn']))
(1, 1, 'S')