Create Project 3 (part 3)
The aim of this exercise is to extend our simple robot language from Question 1 by supporting a simple form of iteration. Note that your solution to this question does not need to support macros from Question 2.

Our robot language now should include iteration statements of the form 'repeat' count body 'end', where the iteration statement starts with the string 'repeat' and is terminated with the string 'end'. Following the 'repeat' keyword is an integer greater than zero, which specifies a count how many times the iteration should be performed. Between the integer count and the 'end' keyword is the body of the iteration, which is the sequence of one or more statements that that will be repeated the given count number of times.

For example, we can use the repeat statement to write complex programs in a more compact form:

the program ['repeat', 3, 'move', 'end'] is equivalent to the program ['move', 'move', 'move'], and would evaluate to (0, 3, 'N')
the program ['repeat', 2, 'move', 'turn', 'end'] is equivalent to the program ['move', 'turn', 'move', 'turn'], and would evaluate to (1, 1, 'S')
the program ['repeat', 2, 'repeat', 2, 'move', 'turn', 'end', 'end'] is equivalent to the program ['move', 'turn', 'move', 'turn', 'move', 'turn', 'move', 'turn'], and would evaluate to (0, 0, 'N').
Note that as shown in the third example above, the body of a repeat statement can contain any number and combination of 'move', 'turn' and 'repeat' statements. In other words, our language can support nested repeat statements up to an arbitrary depth. Our language can also support multiple sequential repeat statements inside another repeat statement, e.g., ['repeat', 2, 'repeat', 2, 'turn', 'end', 'move', 'repeat', 2, 'turn', 'end', 'end'] is a valid program, and is equivalent to ['turn', 'turn', 'move', 'turn', 'turn', 'turn', 'turn', 'move', 'turn', 'turn']. Note that an 'end' statement should be matched with the nearest 'repeat' statement.

You can assume that you will be given a valid example of this type of program. You will not be given a list containing any invalid operators, or any syntactically incorrect repeat statements. The robot always starts at grid coordinates .

You should write a function repeat_interpreter(code) that takes a list code containing a program that comprises 'move', 'turn' and 'repeat' statements, and returns a tuple corresponding to the new state of the robot after executing the program.

Note that it is quite tricky to implement nested repeat statements. If you can only implement an interpreter that supports a single level of repeat statements rather than nested repeat statements, then you should submit that and you are likely to get at least some marks for this question.

Example calls to the function are:

>>> print(repeat_interpreter(['repeat', 3, 'move', 'end']))
(0, 3, 'N')
>>> print(repeat_interpreter(['repeat', 2, 'move', 'turn', 'end']))
(1, 1, 'S')
>>> print(repeat_interpreter(['turn', 'repeat', 3, 'move', 'end']))
(3, 0, 'E')
>>> print(repeat_interpreter(['repeat', 2, 'repeat', 2, 'move', 'turn', 'end', 'end']))
(0, 0, 'N')