Update
The aim of this exercise is to extend our simple robot language, so that we can predefine a library of simple functions called macros.

A macro is a list containing a sequence of operators, where the operators can be 'move', 'turn' or the name of another macro, given as a string. The name of a macro is given as a string, and the body of the macro (i.e., the list of operators) is associated with this name using a dictionary of macros. The advantage of using macros is that they help us to shorten our programs. For example, the dictionary {'turnleft' : ['turn', 'turn', 'turn'], 'turnright' : ['turn'], 'bigleftturn' : ['move', 'move', 'turnleft', 'move', 'move'], 'bigrightturn' : ['move', 'move', 'turnright', 'move', 'move']} contains four macros. We can use these macros to write new programs, such as:

the program ['turnleft', 'turnright'] would evaluate to (0, 0, 'N')
the program ['bigleftturn', 'bigrightturn'] would evaluate to (-4, 4, 'N')
the equation ['bigrightturn', 'turn'] would evaluate to (2, 2, 'S').
You can assume that you will be given a valid example of this type of program. You will not be given a list containing any invalid operators, or any macros that have not been defined in the given dictionary. You can also assume that the macros will not result in any infinite loops, i.e., no macros are defined in terms of themselves, either directly or indirectly. The robot always starts at grid coordinates .

You should write a function macro_interpreter(code, macros) that takes a list code containing the operators 'move' or 'turn' or any macros defined in the dictionary macros, and returns a tuple corresponding to the new state of the robot after executing the program.

Example calls to the function are:


>>> print(macro_interpreter(['turnleft'], {'turnleft' : ['turn', 'turn', 'turn'], 'turnright' : ['turn'], 'bigleftturn' : ['move', 'move', 'turnleft', 'move', 'move'], 'bigrightturn' : ['move', 'move', 'turnright', 'move', 'move']}))
(0, 0, 'W')
>>> print(macro_interpreter(['turnleft', 'turnright'], {'turnleft' : ['turn', 'turn', 'turn'], 'turnright' : ['turn'], 'bigleftturn' : ['move', 'move', 'turnleft', 'move', 'move'], 'bigrightturn' : ['move', 'move', 'turnright', 'move', 'move']}))
(0, 0, 'N')
>>> print(macro_interpreter(['bigleftturn', 'bigrightturn'], {'turnleft' : ['turn', 'turn', 'turn'], 'turnright' : ['turn'], 'bigleftturn' : ['move', 'move', 'turnleft', 'move', 'move'], 'bigrightturn' : ['move', 'move', 'turnright', 'move', 'move']}))
(-4, 4, 'N')
>>> print(macro_interpreter(['circle'], {'turnleft' : ['turn', 'turn', 'turn'], 'turnright' : ['turn'], 'circle' : ['turnleft', 'turn']}))
(0, 0, 'N')