Create Project 1 (part 1)
The aim of this exercise is to implement a function for finding when the trajectory of an object is closest to a given point. A point is given by a tuple as a pair of  coordinates. A trajectory of an object is given by a list of tuples, where each tuple  corresponds to the observed position of the object at a particular time . Your goal is to write a function that returns the time when the object was observed to be closest to the given point.

For example, consider the trajectory , which contains four observations of the position of an object at times , ,  and . If we compare this trajectory to the point , then the closest observation in the trajectory occurs at time .

Write a function closestTime(tr, pt) that takes a list argument tr that corresponds to the trajectory of an object, and a tuple argument pt that corresponds to a test point, and returns the int value of the time of the earliest observation in tr that is closest to the test point pt. Note that if two or more observations are closest to the test point (i.e., they are equally close to the test point, and closer than any other observations), then you should return the time of the earliest such observation.

You can assume that the given trajectory contains at least one observation, and that the  and  coordinates in the observation are float values, and the time  of the observation is a non-negative int. You can also assume that the  and  coordinates of the test point are float values. You can assume that the observations in the trajectory are ordered by increasing time values.

You should calculate the distance between an observation and the test point using the Euclidean distance between two points  and :
Example calls to the function are:


>>> print(closestTime([(0,0,0),(1,1,1),(2,2,2),(1,1,3)], (2,2)))
2
>>> print(closestTime([(0,0,0),(1,1,1),(2,2,2),(1,1,3)], (1.2,1.1)))
1
>>> print(closestTime([(0,0,2),(1,1,3),(2,2,4),(1,1,5)], (-2,-2)))
2
>>> print(closestTime([(0,0,0)], (-2,-2)))
0