Create Project 1 (part 3)
The next trajectory analysis function that you will implement compares two trajectories to see whether two objects are moving together. We can do this by comparing observations that occur at the same time in each trajectory, and checking if their distance is less than or equal to a given threshold. If so, we consider those observations to be close. The aim is to find the maximum number of consecutive observations in the two trajectories, such that each pair of observations occurred at the same time and were close together.

Write a function together(tr1, tr2, th) that takes two list arguments tr1, tr2 (each corresponding to a trajectory), and a float argument th (corresponding to the threshold used to test if two observations are close), and returns the int count of the maximum number of consecutive observations in each trajectory such that the observations occur at the same time and are close.

You can assume that the list arguments tr1, tr2 each contain at least one observation, that the observations in each trajectory are ordered by increasing time values, and that there are no missing time values between the start and end times of a trajectory. Note that the given trajectories tr1, tr2 can have different start and/or end times. You can also assume that the threshold th is greater than zero. Note that you will need to calculate the Euclidean distance between two points in order to determine if they are close.

Example calls to the function are:


>>> print(together([(0,0,0),(1,1,1),(2,2,2),(1,1,3),(1,1,4)], [(0,0,0),(6,1,1),(2,4,2),(1,1,3),(1,0.7,4)], 0.5))
2
>>> print(together([(0,0,0),(1,1,1),(2,2,2),(1,1,3),(1,1,4)], [(0,0,0),(6,1,1),(2,4,2),(1,1,3),(1,0.7,4)], 0.1))
1
>>> print(together([(0,0,2),(1,1,3),(2,2,4),(1,1,5)], [(0,0,0),(1,1,1),(2,2,2)], 0.5))
0
>>> print(together([(0,0,0)], [(0,0,1)], 0.5))
0