Create Project 2 (part 1)
The aim of this exercise is to evaluate the similarity of two people based on what restaurants they like or dislike.

Each person is represented by a tuple that contains the restaurant ratings of that person. Each rating value is an integer between 0 to 5 (inclusive), where 5 is the most favourable rating, 1 is the least favourable rating, and 0 means the person has not rated the restaurant. The number of rating values in each person's rating tuple is the same; i.e., the first rating value will be the rating for restaurant 1, the second rating value will be the rating for restaurant 2, and so on.

For example, the tuple:


((5, 4, 2, 1), (1, 2, 5, 5), (4, 5, 0, 0))
specifies the ratings of three people for four restaurants. The first person liked the first and second restaurants, but didn't like the third and fourth restaurants. In contrast, the second person didn't like the first and second restaurants, but liked the third and fourth restaurants. The third person liked the first and second restaurants, but hasn't rated the third and fourth restaurants.

You can assume that all inputs are of the correct type, that all ratings are between 0 to 5 (inclusive), and that the number of rating values in each person's tuple is the same; i.e., the tuples of ratings all have the same length. You can also assume that the number of restaurants will be greater than zero, and that each person has at least one non-zero rating in their tuple of ratings.

Your goal is to write a function that calculates the similarity between two people based on their respective tuple of restaurant ratings. We will use a measure of similarity called the cosine similarity. Let the tuple of ratings for the first person (with index 0) be , where  is the number of restaurants. Similarly, let the tuple of ratings for the second person (with index 1) be . Let  denote the set of restaurants that have been given a non-zero rating by both people. Given this, we can define the cosine similarity between the person  and person  as:
where  means “take the sum of  using all values in the set .”

For example, the cosine similarity between the first and second example people above would be:
The cosine similarity between the first and third example people above would be:
You should write a function similarity(p1, p2, ratings) that takes the index p1 of one person in ratings, the index p2 of another person in ratings, and a tuple of tuples ratings corresponding to the ratings of a set of people, and returns the float value of the cosine similarity of the two specified people in terms of their ratings. If there are no restaurants that have been given non-zero ratings by both p1 and p2, then their similarity is 0.0.

Note: The marker in Grok will automatically round the return value from your function to two decimal places before marking. As such, for the first example above, any number you return that rounds to 0.56 when rounding to two decimal places will be marked as correct.

Here are some example calls to your similarity function:


>>> print(similarity(0, 1, ((5, 4, 2, 1), (1, 2, 5, 5), (4, 5, 0, 0))))
0.556669940741281
>>> print(similarity(0, 2, ((5, 4, 2, 1), (1, 2, 5, 5), (4, 5, 0, 0))))
0.975609756097561
>>> print(similarity(0, 0, ((5, 4, 2, 1), (1, 2, 5, 5), (4, 5, 0, 0))))
1.0
>>> print(similarity(1, 2, ((5, 4, 2, 1), (0, 0, 5, 5), (4, 5, 0, 0))))
0.0