Create Project 2 (part 2)
The aim of this exercise is to find the average rating by customers for a given type of restaurant cuisine.

You are given the type of cuisine, a collection of restaurants and their cuisines, and a tuple of tuples of ratings in the same format as the prior question. Your goal is to write a function that calculates the average rating for the given cuisine by averaging all the non-zero ratings by customers for restaurants with that type of cuisine.

The cuisine will be provided as a string; e.g. "italian". Each restaurant will be represented by a tuple of two elements: the name of the restaurant as a string, and the type of cuisine in that restaurant as a string. For example, ("pizza palace","italian") specifies an Italian restaurant called "pizza palace". You will be given a collection of restaurants in the form of a tuple of such tuples. You will also be given the ratings of a number of people. This will have the same format as the ratings argument in Question 1, i.e., a tuple of the ratings from different people, where each person's ratings are given by a tuple of integer rating values for each restaurant in the given collection of restaurants. Like Question 1, each rating is an integer between 0 to 5, where 5 is the most favourable rating, 1 is the least favourable rating, and 0 means the person has not rated the restaurant.

For example, consider a collection of six restaurants given as:


(("top thai","thai"), ("pizza palace","italian"), ("krabi hut","thai"), ("tacky thai","thai"), ("papa","italian"), ("texmex","mexican"))
We are given a collection of ratings from seven people:


((5, 1, 5, 4, 1, 2), (4, 1, 0, 5, 2, 0), (1, 5, 1, 2, 5, 4), (0, 5, 1, 1, 0, 5), (1, 5, 0, 1, 5, 0), (5, 1, 5, 0, 1, 0), (5, 1, 4, 5, 1, 0))
The average rating for the cuisine "mexican" will be . Note that four of the seven people have not rated the one Mexican restaurant, so the average is calculated based on the three ratings given.

You should write a function ave_rating(cuisine, restaurants, ratings) that takes a string cuisine, a tuple of restaurants, and a tuple ratings containing the ratings from a collection of people, and returns a float corresponding to the average of the non-zero ratings given by all people to all restaurants of the given cuisine. You can assume that all inputs are of the correct type. You can assume that the tuples of ratings from each person all have the same length, which is equal to the number of restaurants in the restaurants tuple. You can also assume that the order of ratings in the tuple of ratings from each person is consistent with the order of the restaurants in the restaurants tuple.

If no one has rated a restaurant of the given cuisine, your function should return 0. Similarly, if there are no restaurants of the given cuisine, your function should return 0. Like Question 1, the marker in Grok will automatically round the return value from your function to two decimal places before marking.

Example calls to the function are:


>>> print(ave_rating("thai", (("top thai","thai"), ("pizza palace","italian"), ("krabi hut","thai"), ("tacky thai","thai")), ((5, 4, 2, 1), (1, 2, 5, 5), (4, 5, 0, 0))))
3.2857142857142856
>>> print(ave_rating("italian", (("top thai","thai"), ("pizza palace","italian"), ("krabi hut","thai"), ("tacky thai","thai")), ((5, 4, 2, 1), (1, 2, 5, 5), (4, 5, 0, 0))))
3.6666666666666665
>>> print(ave_rating("vulcan", (("top thai","thai"), ("pizza palace","italian"), ("krabi hut","thai"), ("tacky thai","thai")), ((5, 4, 2, 1), (1, 2, 5, 5), (4, 5, 0, 0))))
0
>>> print(ave_rating("american", (("mcdude","american"),), ((1,), (1,), (2,), (1,))))
1.25